relocated function definitions
